Fixed PortfolioOverview/ViewFormPanel issues with resources
