+++ SharePointFramework/PortfolioWebParts/src/components/PortfolioOverview/ViewFormPanel/index.tsx
@@ -13,7 +13,7 @@ export const ViewFormPanel: FC = () => {
   return (
     <Panel
       isOpen={context.state.viewForm.isOpen}
-      headerText={isEditing ? strings.EditColumnHeaderText : strings.NewColumnHeaderText}
+      headerText={isEditing ? strings.EditViewHeaderText : strings.NewViewHeaderText}
       onRenderFooterContent={() => <ViewFormPanelFooter onSave={onSave} />}
       isFooterAtBottom={true}
       onDismiss={onDismiss}
@@ -57,6 +57,7 @@ export const ViewFormPanel: FC = () => {
           required={true}
           value={view.get('iconName')}
           onChange={(_, value) => setView('iconName', value)}
+          iconProps={{ iconName: view.get('iconName') }}
         />
       </FormFieldContainer>
       <FormFieldContainer description={strings.DefaultViewDescription}>
